utils: Move find_distro_path() to distro.py
